third_party/WebKit/LayoutTests/svg/transforms/transform-boxsize-usecounter-no-trigger-2.html

<!DOCTYPE html>
<title>SVG transform: use counter does not trigger for translations</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style>
rect {
  transform-origin: 50% 50%;
  transform: translate(100, 100);
}
</style>
<svg>
  <rect width="100" height="100" fill="green"/>
</svg>
<script>
// From UseCounter.h
const TransformUsesBoxSizeOnSVG = 1842;

test(function() {
  document.querySelector('rect').getBoundingClientRect();
  assert_false(internals.isUseCounted(document, TransformUsesBoxSizeOnSVG));
});
</script>
